requirements.txt
django==4.0.7
celery==5.2.7
redis==4.3.4
flower==1.2.0
mariadb==1.1.6
mysqlclient==2.1.1
djangorestframework==3.14.0
django-oauth-toolkit==2.2.0